PIM kernel module basics
The kernel module is in the 'module' subdirectory. It includes the
standard interface for a kernel module as well as the frontswap API.
There are files for modeling UPMEM's DPU (roughly) as well as parts
of their userspace API. I chose the userspace API because I'm most
familiar with it, it is published, and our snappy implementation
uses it.
I have included the snappy compressor, which includes the sequential
reader as well. I have borrowed a few files from UPMEM (BSD license)
to keep things as standard as possible (and minimize work).
The config directory contains the changes needed to buildroot for
testing.
The setup.sh script is meant to build the complete environment on
Ubuntu 20.04 but I don't think its complete yet. The run-qemu.sh
script assumes that everything is built succesfully and the kernel
and initrd files are in the output directory.

Signed-off-by: Joel Nider <joel@nider.org>