Add statistics output
To analyze results & generate graphs, I need to get some output. I write to
the kernel log after each event (store, load, invalidate) some statistics,
mainly to see the % utilization of the storage area.

I also removed a lot of debug messages in order to make the output a
tractable length. There is a bash script to parse the output, but I still
don't want it to take forever. The script generates a CSV from the raw log.

I added a script to generate a graph from the CSV output. This requires
python and pandas and probably a lot of other related packages to generate
the graph.

I fixed 3 serious bugs:
(1) Only overwrite hash entries that belong to the same tag
(2) Don't use the out buffer for invalidation id's
(3) Pages in the out buffer may need to be invalidated

Signed-off-by: Joel Nider <joel@nider.org>