Copy the DPU header files to the correct location
Before each build, these header files containing the kernel API
functions for managing the DPUs, are copied to the correct location.
When the kernel module is built, these header files are used to
provide the API function declarations. Since they are not part of
the upstream kernel, they must be copied before the kernel module
can be built. Currently, this is part of the make file, but there
are other ways to do this (such as patching the kernel source from
buildroot).
This is not meant as the overall solution, but is an important step
to port the DPU simulation code to the real DPUs.

Signed-off-by: Joel Nider <joel@nider.org>